# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Habit Journey API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /journey-rewards:
    get:
      operationId: listJourneyRewards
      tags: [rewards]
      summary: List the current owner's private real-world journey rewards and history
      responses:
        "200":
          description: Owner-only journey rewards, including unattached pending rewards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/JourneyReward" }
    post:
      operationId: createJourneyReward
      tags: [rewards]
      summary: Create a private physical or experience reward, optionally attaching it to an existing journey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JourneyRewardCreateInput" }
      responses:
        "201":
          description: Created pending journey reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JourneyReward" }
        "400":
          description: Invalid reward or attachment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: The selected journey already has a real-world reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /journey-rewards/{journeyRewardId}:
    get:
      operationId: getJourneyReward
      tags: [rewards]
      summary: Read one of the current owner's private real-world journey rewards
      parameters:
        - name: journeyRewardId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Owner's journey reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JourneyReward" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    patch:
      operationId: updateJourneyReward
      tags: [rewards]
      summary: Update a pending journey reward after explicit user confirmation
      parameters:
        - name: journeyRewardId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JourneyRewardUpdateInput" }
      responses:
        "200":
          description: Updated pending journey reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JourneyReward" }
        "400":
          description: Confirmation was not provided or the update is invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Only pending rewards can be updated or the selected journey already has a reward
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /journey-rewards/{journeyRewardId}/claim:
    post:
      operationId: claimJourneyReward
      tags: [rewards]
      summary: Record the owner's claim of an unlocked personal reward without payment or coin spending
      parameters:
        - name: journeyRewardId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Claimed reward; repeated claims return the same claimed state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JourneyReward" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: The reward has not been unlocked by eligible server-side journey completion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    JourneyRewardCreateInput:
      type: object
      additionalProperties: false
      required: [title, type]
      description: Create an owner-private real-world reward, optionally attached to an existing journey. Omit habitId to leave it unattached.
      properties:
        habitId: { type: ["integer", "null"], minimum: 1 }
        title: { type: string, minLength: 1 }
        type: { type: string, enum: [physical, experience] }
        description: { type: string, maxLength: 2000 }
        imageObjectPath:
          type: string
          minLength: 1
          description: Private normalized object path returned by the existing photo upload flow; never a public URL or file bytes.
        estimatedValue: { type: number, minimum: 0 }
    JourneyRewardUpdateInput:
      type: object
      additionalProperties: false
      required: [confirm]
      description: Updates are accepted only while status is pending and only with explicit confirmation=true.
      properties:
        confirm: { type: boolean, enum: [true] }
        habitId: { type: ["integer", "null"], minimum: 1, description: Attach to an existing owner journey or explicitly detach with null. }
        title: { type: string, minLength: 1 }
        type: { type: string, enum: [physical, experience] }
        description: { type: ["string", "null"], maxLength: 2000 }
        imageObjectPath:
          type: ["string", "null"]
          minLength: 1
          description: Private normalized object path returned by the existing photo upload flow; null removes the image.
        estimatedValue: { type: ["number", "null"], minimum: 0 }
    JourneyReward:
      type: object
      required: [id, habitId, title, type, description, imageUrl, estimatedValue, status, createdAt, updatedAt, unlockedAt, claimedAt, currentDay, daysRemaining]
      description: Owner-private personal reward and its server-computed 22-calendar-day progress. Image URL is a normalized private object path.
      properties:
        id: { type: integer }
        habitId: { type: ["integer", "null"] }
        title: { type: string }
        type: { type: string, enum: [physical, experience] }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"], description: Normalized private object path; clients serve it through the authenticated object route. }
        estimatedValue: { type: ["number", "null"], minimum: 0 }
        status: { type: string, enum: [pending, unlocked, claimed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        unlockedAt: { type: ["string", "null"], format: date-time }
        claimedAt: { type: ["string", "null"], format: date-time }
        currentDay: { type: ["integer", "null"], minimum: 0, maximum: 22, description: Current server-computed calendar day of the attached journey; null when unattached. }
        daysRemaining: { type: ["integer", "null"], minimum: 0, maximum: 22, description: Remaining calendar days in the attached 22-day journey; null when unattached. }
...