# lib/api-zod/src/generated/types/journeyReward.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Habit Journey API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JourneyRewardStatus } from './journeyRewardStatus';
import type { JourneyRewardType } from './journeyRewardType';

/**
 * Owner-private personal reward and its server-computed 22-calendar-day progress. Image URL is a normalized private object path.
 */
export interface JourneyReward {
  id: number;
  /** @nullable */
  habitId: number | null;
  title: string;
  type: JourneyRewardType;
  /** @nullable */
  description: string | null;
  /**
     * Normalized private object path; clients serve it through the authenticated object route.
     * @nullable
     */
  imageUrl: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  estimatedValue: number | null;
  status: JourneyRewardStatus;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  unlockedAt: Date | null;
  /** @nullable */
  claimedAt: Date | null;
  /**
     * Current server-computed calendar day of the attached journey; null when unattached.
     * @minimum 0
     * @maximum 22
     * @nullable
     */
  currentDay: number | null;
  /**
     * Remaining calendar days in the attached 22-day journey; null when unattached.
     * @minimum 0
     * @maximum 22
     * @nullable
     */
  daysRemaining: number | null;
}