# lib/api-zod/src/generated/types/journeyRewardCreateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Habit Journey API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JourneyRewardCreateInputType } from './journeyRewardCreateInputType';

/**
 * Create an owner-private real-world reward, optionally attached to an existing journey. Omit habitId to leave it unattached.
 */
export interface JourneyRewardCreateInput {
  /**
     * @minimum 1
     * @nullable
     */
  habitId?: number | null;
  /** @minLength 1 */
  title: string;
  type: JourneyRewardCreateInputType;
  /** @maxLength 2000 */
  description?: string;
  /**
     * Private normalized object path returned by the existing photo upload flow; never a public URL or file bytes.
     * @minLength 1
     */
  imageObjectPath?: string;
  /** @minimum 0 */
  estimatedValue?: number;
}